R wrapper for the interactive_tour function written on python

R wrapper for the interactive_tour function written on python

Usage

interactive_tour(data, col_names, plot_objects, half_range, n_max_cols)

Arguments

data

the dataset you want to investigate

col_names

the column names of your dataset

plot_objects

a named list of objects you want to be displayed. Each entry requires a definition of the type of display and a specification of what should be plotted. For tours, a history object, and for histograms and scatterplots, the names of the variables to be displayed have to be provided (see example).

half_range

can be specified directly or be calculated done in the tourr package.

n_max_cols

specifies the maximal number of columns of the grid of the final display.

Value

Examples

library(tourr)
library(devtools)
#> Error in library(devtools): there is no package called ‘devtools’
library(reticulate)
library(pytourr)
f <- apply(flea[,1:6], 2, function(x) (x-mean(x))/sd(x))
guided_tour_history <- save_history(f,
                                   tour_path = guided_tour(holes()))
#> Value  0.993   15.8 % better  - NEW BASIS
#> Value  1.080   8.8 % better  - NEW BASIS
#> Value  1.097   1.6 % better  - NEW BASIS
#> Value  1.119   1.9 % better  - NEW BASIS
#> Value  1.121   0.2 % better  - NEW BASIS
#> Value  1.125   0.3 % better  - NEW BASIS
#> Value  1.132   0.6 % better  - NEW BASIS
#> Value  1.144   1.1 % better  - NEW BASIS
#> Value  1.147   0.2 % better  - NEW BASIS
#> Value  1.150   0.3 % better  - NEW BASIS
#> Value  1.151   0.1 % better  - NEW BASIS
#> Value  1.153   0.1 % better  - NEW BASIS
#> Value  1.155   0.2 % better  - NEW BASIS
#> Value  1.156   0.1 % better  - NEW BASIS
#> Value  1.160   0.4 % better  - NEW BASIS
#> Value  1.163   0.3 % better  - NEW BASIS
#> Value  1.164   0.0 % better 
#> Value  1.166   0.2 % better  - NEW BASIS
#> Value  1.191   2.1 % better  - NEW BASIS
#> Value  1.200   0.8 % better  - NEW BASIS
#> Value  1.211   0.9 % better  - NEW BASIS
#> Value  1.220   0.7 % better  - NEW BASIS
#> Value  1.241   1.7 % better  - NEW BASIS
#> Value  1.248   0.6 % better  - NEW BASIS
#> Value  1.254   0.4 % better  - NEW BASIS
#> Value  1.259   0.4 % better  - NEW BASIS
#> Value  1.266   0.5 % better  - NEW BASIS
#> Value  1.270   0.3 % better  - NEW BASIS
#> Value  1.272   0.2 % better  - NEW BASIS
#> Value  1.274   0.1 % better  - NEW BASIS
#> Value  1.274   0.0 % better 
#> Value  1.275   0.1 % better 
#> Value  1.275   0.1 % better 
#> Value  1.275   0.1 % better 
#> Value  1.274   0.0 % better 
#> Value  1.274   0.0 % better 
#> Value  1.274   0.0 % better 
#> Value  1.274   0.1 % better 
#> Value  1.275   0.1 % better 
#> Value  1.274   0.0 % better 
#> Value  1.274   0.0 % better 
#> Value  1.274   0.0 % better 
#> Value  1.275   0.1 % better 
#> Value  1.275   0.1 % better  - NEW BASIS
#> Value  1.276   0.0 % better 
#> Value  1.277   0.1 % better  - NEW BASIS
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.1 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.1 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.278   0.1 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> Value  1.277   0.0 % better 
#> No better bases found after 25 tries.  Giving up.
#> Final projection: 
#> 0.326  -0.602  
#> 0.566  0.026  
#> 0.335  0.167  
#> 0.511  0.355  
#> 0.342  -0.538  
#> 0.290  0.441  
grand_tour_history <- save_history(f,
                                  tour_path = grand_tour())
tour_history_1d <- save_history(f,
                               tour_path = grand_tour(d=1))

half_range <- max(sqrt(rowSums(f^2)))
col_names <- colnames(f)

init_env()
#> Python environment "r-pytourr" successfully loaded 

obj1 <- list(type = "2d_tour", obj = guided_tour_history)
obj2 <- list(type = "1d_tour", obj = tour_history_1d)
obj3 <- list(type = "scatter", obj = c("tars1", "tars2"))
obj4 <- list(type = "hist", obj = "head")

interactive_tour(f,
                col_names,
               list(obj1,obj2,obj3,obj4),
                half_range,
                n_max_cols=2)
#> Error in py_run_file_impl(file, local, convert): ModuleNotFoundError: No module named '_tkinter'
#> Run `reticulate::py_last_error()` for details.